flasky/requirements.txt
alembic==1.0.11
bleach==3.1.0
blinker==1.4
Click==7.0
dominate==2.3.5
entrypoints==0.3
Faker==2.0.0
flake8==3.7.8
Flask==1.1.1
Flask-Avatars==0.2.2
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.4.3
Flask-Dropzone==1.5.4
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-SocketIO==4.1.0
Flask-SQLAlchemy==2.4.0
Flask-WhooshAlchemyPlus==0.7.5
flask-whooshee==0.7.0
Flask-WTF==0.14.2
gevent==1.4.0
gevent-websocket==0.10.1
greenlet==0.4.15
itsdangerous==1.1.0
jieba==0.39
Jinja2==2.10.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
paginate==0.5.6
Pillow==6.1.0
pkg-resources==0.0.0
pycodestyle==2.5.0
pyflakes==2.1.1
PyMySQL==0.9.3
python-dateutil==2.8.0
python-editor==1.0.4
python-engineio==3.8.2.post1
python-socketio==4.2.1
six==1.12.0
SQLAlchemy==1.3.6
text-unidecode==1.2
visitor==0.1.3
webencodings==0.5.1
Werkzeug==0.15.5
Whoosh==2.7.4
WTForms==2.2.1
